backend/app/ocr_processor_minimal.py
import os
import logging
from PIL import Image

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class OCRProcessor:
    def __init__(self):
        logger.info("Initializing Minimal OCR processor...")
        self.ocr_available = True
        self.temp_dir = "temp_ocr"
        os.makedirs(self.temp_dir, exist_ok=True)
        logger.info("Minimal OCR processor initialized!")
    
    def process_document(self, file_path):
        logger.info(f"Processing document: {file_path}")
        
        if not os.path.exists(file_path):
            logger.error(f"File not found: {file_path}")
            return {'text': '', 'confidence': 0, 'word_count': 0, 'bounding_boxes': []}
        
        file_ext = os.path.splitext(file_path)[1].lower()
        file_size = os.path.getsize(file_path)
        
        try:
            if file_ext == '.pdf':
                return self._process_pdf(file_path, file_size)
            elif file_ext in ['.jpg', '.jpeg', '.png', '.tiff', '.bmp', '.webp']:
                return self._process_image(file_path, file_size)
            else:
                logger.warning(f"Unsupported file type: {file_ext}")
                return {'text': 'Unsupported file format', 'confidence': 0, 'word_count': 0, 'bounding_boxes': []}
        except Exception as e:
            logger.error(f"Document processing failed: {e}")
            return {'text': f'Processing failed: {str(e)}', 'confidence': 0, 'word_count': 0, 'bounding_boxes': []}
    
    def _process_pdf(self, file_path, file_size):
        logger.info(f"Processing PDF: {file_path}")
        
        try:
            filename = os.path.basename(file_path)
            placeholder_text = f"""PDF Document: {filename}

This PDF document has been successfully uploaded and is ready for processing.

Document Information:
- File Name: {filename}
- File Size: {file_size:,} bytes ({file_size/1024:.1f} KB)
- File Type: PDF Document
- Status: Successfully loaded

Text Extraction Status:
Currently using minimal processor. For full text extraction from PDFs, 
the system needs PyMuPDF (fitz) library to be properly installed.

To enable full PDF processing:
1. Fix PyMuPDF installation issues
2. Or use an external OCR service
3. Or convert PDF to images first

The document structure and metadata can be processed, but text extraction 
requires additional OCR capabilities to be configured."""

            word_count = len(placeholder_text.split())
            logger.info(f"PDF placeholder created: {word_count} words")
            
            return {
                'text': placeholder_text,
                'confidence': 0.3,
                'word_count': word_count,
                'bounding_boxes': [],
                'file_info': {
                    'filename': filename,
                    'size': file_size,
                    'type': 'pdf'
                },
                'method': 'minimal_pdf_info'
            }
        except Exception as e:
            logger.error(f"PDF processing failed: {e}")
            return {'text': f'PDF processing failed: {str(e)}', 'confidence': 0, 'word_count': 0, 'bounding_boxes': []}
    
    def _process_image(self, file_path, file_size):
        logger.info(f"Processing image: {file_path}")
        
        try:
            image = Image.open(file_path)
            width, height = image.size
            mode = image.mode
            filename = os.path.basename(file_path)
            
            placeholder_text = f"""Image Document: {filename}

This image has been successfully uploaded and analyzed.

Image Information:
- File Name: {filename}
- Dimensions: {width} x {height} pixels
- Color Mode: {mode}
- File Size: {file_size:,} bytes ({file_size/1024:.1f} KB)
- Aspect Ratio: {width/height:.2f}

OCR Analysis Status:
Image loaded successfully and is ready for text extraction.
Currently using minimal processor for demonstration.

For Text Extraction:
To extract actual text from this image, install one of these OCR engines:

1. EasyOCR:
   pip install easyocr

2. Tesseract OCR:
   - Install Tesseract software
   - pip install pytesseract

3. Cloud OCR Services:
   - Google Vision API
   - Amazon Textract
   - Azure Cognitive Services

Image Quality Assessment:
- Resolution: {"High" if width * height > 1000000 else "Medium" if width * height > 100000 else "Low"}
- Size Category: {"Large" if file_size > 1000000 else "Medium" if file_size > 100000 else "Small"}
- Suitable for OCR: {"Yes" if width > 300 and height > 300 else "May need enhancement"}"""

            word_count = len(placeholder_text.split())
            logger.info(f"Image analyzed: {width}x{height} {mode} image")
            
            return {
                'text': placeholder_text,
                'confidence': 0.5,
                'word_count': word_count,
                'bounding_boxes': [],
                'image_info': {
                    'filename': filename,
                    'width': width,
                    'height': height,
                    'mode': mode,
                    'file_size': file_size,
                    'resolution': width * height
                },
                'method': 'minimal_image_analysis'
            }
        except Exception as e:
            logger.error(f"Image processing failed: {e}")
            return {'text': f'Image processing failed: {str(e)}', 'confidence': 0, 'word_count': 0, 'bounding_boxes': []}
    
    def extract_layout_elements(self, file_path):
        try:
            result = self.process_document(file_path)
            return {
                'text_blocks': [],
                'word_count': result.get('word_count', 0),
                'confidence': result.get('confidence', 0),
                'method': 'minimal',
                'file_info': result.get('file_info', {}) or result.get('image_info', {})
            }
        except Exception as e:
            logger.error(f"Layout extraction failed: {e}")
            return {'text_blocks': [], 'word_count': 0, 'confidence': 0, 'error': str(e)}

    @property
    def tesseract_available(self):
        return False
