backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
pytesseract==0.3.10
opencv-python==4.8.1.78
pillow==10.1.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
pdfplumber==0.10.3
spacy==3.7.2
easyocr>=1.7.0
opencv-python>=4.8.0
torch>=1.13.0
torchvision>=0.14.0
